Refactor resource tray, fix css
