+++ less/layout.less
@@ -45,7 +45,6 @@
 
 	.equipment {
 		width: 25%;
-		padding: 10px;
 		display: inline-flex;
 		flex-direction: column;
 		justify-content: space-between;
@@ -55,6 +54,7 @@
 			max-width: 100%;
 			height: 80%;
 			display: flex;
+			padding: 10px 10px 0px 10px;
 
 			> section {
 				height: 100%;
@@ -134,8 +134,12 @@
 }
 
 .tray {
+	border-top: 2px #b5b58d solid;
 	display: block;
-	width: 80%;
+	width: 100%;
+	> span {
+		border-right: 1px #b5b58d solid;
+	}
 }
 
 @secret-size: 20%;
